+++ R/plotsNtables.R
@@ -11,9 +11,10 @@
 #'                 `Sharpe ratio`, `max drawdown`, `annual return`, `annual volatility`, 
 #'                 `Sterling ratio`, `Omega ratio`, `ROT bps`, etc.
 #' @param caption Table caption (only works for \code{type = "DT"}).
-#' @param type Type of table. Valid options: \code{"simple", "DT", "grid.table"}. Default is 
+#' @param type Type of table. Valid options: \code{"simple", "DT", "kable", "grid.table"}. Default is 
 #'             \code{"simple"} and generates a simple matrix (with the other choices the 
 #'             corresponding package must be installed).
+#' @param digits Integer indicating the number of decimal places when rounding (default is 2).
 #' @param order_col Column number or column name of the performance measure to be used to 
 #'                  sort the rows (only used for table \code{type = "DT"}). By default the 
 #'                  last column will be used.
@@ -45,23 +46,30 @@
 #' # do backtest
 #' bt <- portfolioBacktest(list("Quintile" = quintile_portfolio), 
 #'                         dataset10,
-#'                         benchmark = c("uniform", "index"))
+#'                         benchmark = c("1/N", "index"))
 #' 
 #' # now we can obtain the table
 #' bt_summary_median <- backtestSummary(bt)
 #' summaryTable(bt_summary_median, measures = c("max drawdown", "annual volatility"))
 #' summaryTable(bt_summary_median, measures = c("max drawdown", "annual volatility"), type = "DT")
+#' summaryTable(bt_summary_median, type = "kable") 
+#' # this returned kable object can be combined with: " |> kableExtra::kable_styling()"
 #' }
 #' 
 #' @export
 summaryTable <- function(bt_summary, measures = NULL, caption = "Performance table",
-                         type = c("simple", "DT", "grid.table"), 
+                         type = c("simple", "DT", "kable", "grid.table"), 
+                         digits = 2,
                          order_col = NULL, order_dir = c("asc", "desc"), page_length = 10) {
   if (is.null(measures)) measures <- rownames(bt_summary$performance_summary)  # by default use all
   # extract performance measures
   real_measures <- intersect(measures, rownames(bt_summary$performance_summary))
   performance <- bt_summary$performance_summary[real_measures, , drop = FALSE]
-  performance <- t(round(performance, 4))
+  performance <- t(round(performance, digits))
+  
+  # percentage columns for formatting in DT and kable
+  cols_percentage <- intersect(c("annual return", "annual volatility", "max drawdown", "VaR (0.95)", "CVaR (0.95)"),
+                               colnames(performance))
   
   # show table
   switch(match.arg(type),
@@ -72,21 +80,36 @@ summaryTable <- function(bt_summary, measures = NULL, caption = "Performance tab
            if (is.character(order_col)) order_col <- which(colnames(performance) == order_col)
            if (is.null(order_col) || length(order_col) == 0) order_col <- ncol(performance)
            order_dir <- match.arg(order_dir)
-           p <- DT::datatable(performance, 
+           p <- DT::datatable(performance,
                               options = list(pageLength = page_length, scrollX = TRUE, order = list(order_col, order_dir)),
                               caption = caption)
-           p <- DT::formatStyle(p, 0, target = "row", fontWeight = DT::styleEqual(c("uniform", "index"), c("bold", "bold")))
-           p <- DT::formatRound(p, colnames(performance), digits = 2)
-           if ("annual volatility" %in% colnames(performance))
-             p <- DT::formatPercentage(p, "annual volatility", 1)
-           if ("max drawdown" %in% colnames(performance))
-             p <- DT::formatPercentage(p, "max drawdown", 1)
+           p <- DT::formatStyle(p, 0, target = "row", fontWeight = DT::styleEqual(c("1/N", "index"), c("bold", "bold")))
+           # rounding
+           p <- DT::formatRound(p, colnames(performance), digits = digits)
            if ("ROT (bps)" %in% colnames(performance))
              p <- DT::formatRound(p, "ROT (bps)", digits = 0)
            if ("cpu time" %in% colnames(performance))
              p <- DT::formatRound(p, "cpu time", digits = 4)
+           if (length(cols_percentage) > 0)
+             p <- DT::formatPercentage(p, cols_percentage, digits = 1)
            p
          },
+         "kable" = {
+           if (!requireNamespace("knitr", quietly = TRUE)) 
+             stop("Please install package \"knitr\" or choose another table type", call. = FALSE)
+           if (!requireNamespace("scales", quietly = TRUE)) 
+             stop("Please install package \"scales\" or choose another table type", call. = FALSE)
+           # data.frame with percentages
+           df <- data.frame("Portfolio" = rownames(performance), 
+                            performance,
+                            check.names = FALSE)
+           df[cols_percentage] <- lapply(df[cols_percentage], 
+                                         FUN = scales::percent, accuracy = 1)
+           # kable
+           knitr::kable(df, digits = digits, booktabs = TRUE, linesep = "", row.names = FALSE, 
+                        align = c('l', rep('r', ncol(df) - 1)), 
+                        caption = caption)
+         },         
          "grid.table" = {
            if (!requireNamespace("gridExtra", quietly = TRUE)) 
              stop("Please install package \"gridExtra\" or choose another table type", call. = FALSE)
@@ -94,6 +117,8 @@ summaryTable <- function(bt_summary, measures = NULL, caption = "Performance tab
            },
          stop("Table type unknown."))
 }
+
+
 
 
 
@@ -137,7 +162,7 @@ summaryTable <- function(bt_summary, measures = NULL, caption = "Performance tab
 #' 
 #' # do backtest
 #' bt <- portfolioBacktest(list("Quintile" = quintile_portfolio), dataset10,
-#'                         benchmark = c("uniform", "index"))
+#'                         benchmark = c("1/N", "index"))
 #'                         
 #' # now we can obtain the table
 #' bt_summary_median <- backtestSummary(bt)
@@ -201,6 +226,7 @@ summaryBarPlot <- function(bt_summary, measures = NULL, type = c("ggplot2", "sim
 #'                 \code{"Sharpe ratio"}, \code{"max drawdown"}, \code{"annual return"}, \code{"annual volatility"}, 
 #'                 \code{"Sterling ratio"}, \code{"Omega ratio"}, and \code{"ROT bps"}.
 #'                  Default is \code{"Sharpe ratio"}.
+#' @param ref_portfolio Reference portfolio (whose measure will be subtracted). Default is \code{NULL}.
 #' @param type Type of plot. Valid options: \code{"ggplot2", "simple"}. Default is 
 #'             \code{"ggplot2"}.
 #' @param ... Additional parameters. For example: 
@@ -230,7 +256,7 @@ summaryBarPlot <- function(bt_summary, measures = NULL, type = c("ggplot2", "sim
 #' 
 #' # do backtest
 #' bt <- portfolioBacktest(list("Quintile" = quintile_portfolio), dataset10,
-#'                         benchmark = c("uniform", "index"))
+#'                         benchmark = c("1/N", "index"))
 #' 
 #' # now we can plot
 #' backtestBoxPlot(bt, "Sharpe ratio")
@@ -243,13 +269,21 @@ summaryBarPlot <- function(bt_summary, measures = NULL, type = c("ggplot2", "sim
 #' @importFrom ggplot2 ggplot aes aes_string geom_boxplot geom_point scale_x_discrete coord_flip labs
 #' @importFrom rlang .data
 #' @export
-backtestBoxPlot <- function(bt, measure = "Sharpe ratio", type = c("ggplot2", "simple"), ...) {
+backtestBoxPlot <- function(bt, measure = "Sharpe ratio", ref_portfolio = NULL, type = c("ggplot2", "simple"), ...) {
   # extract correct performance measure
   res_list_table <- backtestTable(bt)
   # idx <- grep(measure, names(res_list_table), ignore.case = TRUE)  # it does not work when "measure" contains brackets
   idx <- which(measure == names(res_list_table))
   if (length(idx)!=1) stop(measure, " does not match a single performance measure")
   res_table <- res_list_table[[idx]]
+  
+  if (!is.null(ref_portfolio)) {
+    if (!ref_portfolio %in% colnames(res_table))
+      stop("Reference portfolio does not exist.")
+    res_table <- res_table - res_table[, ref_portfolio]
+    #res_table <- res_table[, !colnames(res_table) %in% ref_portfolio, drop = FALSE]
+    measure <- paste0("Excess ", measure, " (w.r.t. reference portfolio ", ref_portfolio, ")")
+  }
   
   # plot boxplot
   params <- list(res_table[, ncol(res_table):1], ...)
@@ -307,7 +341,7 @@ backtestBoxPlot <- function(bt, measure = "Sharpe ratio", type = c("ggplot2", "s
 #' @author Daniel P. Palomar and Rui Zhou
 #' 
 #' @seealso \code{\link{summaryBarPlot}}, \code{\link{backtestBoxPlot}}, 
-#'          \code{\link{backtestChartDrawdown}}, \code{\link{backtestChartStackedBar}}
+#'          \code{\link{backtestChartDrawdown}}, \code{\link{backtestChartStackedBar}}, \code{\link{backtestChartSharpeRatio}}
 #' 
 #' @examples
 #' \donttest{
@@ -326,7 +360,7 @@ backtestBoxPlot <- function(bt, measure = "Sharpe ratio", type = c("ggplot2", "s
 #' 
 #' # do backtest
 #' bt <- portfolioBacktest(list("Quintile" = quintile_portfolio), dataset10,
-#'                         benchmark = c("uniform", "index"))
+#'                         benchmark = c("1/N", "index"))
 #' 
 #' # now we can chart
 #' backtestChartCumReturn(bt)
@@ -377,7 +411,7 @@ backtestChartCumReturn <- function(bt, portfolios = names(bt), dataset_num = 1, 
 #' @author Daniel P. Palomar and Rui Zhou
 #' 
 #' @seealso \code{\link{summaryBarPlot}}, \code{\link{backtestBoxPlot}}, 
-#'          \code{\link{backtestChartCumReturn}}, \code{\link{backtestChartStackedBar}}
+#'          \code{\link{backtestChartCumReturn}}, \code{\link{backtestChartStackedBar}}, \code{\link{backtestChartSharpeRatio}}
 #' 
 #' @examples
 #' \donttest{
@@ -396,7 +430,7 @@ backtestChartCumReturn <- function(bt, portfolios = names(bt), dataset_num = 1, 
 #' 
 #' # do backtest
 #' bt <- portfolioBacktest(list("Quintile" = quintile_portfolio), dataset10,
-#'                         benchmark = c("uniform", "index"))
+#'                         benchmark = c("1/N", "index"))
 #' 
 #' # now we can chart
 #' backtestChartDrawdown(bt)
@@ -434,6 +468,112 @@ backtestChartDrawdown <- function(bt, portfolios = names(bt), dataset_num = 1, t
 
 
 
+#' @title Chart of the rolling Sharpe ratio over time for a single backtest
+#' 
+#' @description Create chart of the rolling Sharpe ratio over time for a single backtest
+#' obtained with the function \code{\link{portfolioBacktest}}.
+#' By default the chart is based on the package \code{ggplot2}, but the user can also 
+#' specify a plot based on \code{PerformanceAnalytics}.
+#' 
+#' @inheritParams backtestChartCumReturn
+#' @param lookback Length of the lookback rolling window in periods (default is \code{100}).
+#' @param by Intervals at which the Sharpe ratio is to be calculated (default is equal to \code{1}).
+#' @param gap Initial number of periods to skip (default is equal to \code{lookback}).
+#' @param bars_per_year Number of bars/periods per year (default is \code{252}).
+#' 
+#' @author Daniel P. Palomar and Rui Zhou
+#' 
+#' @seealso \code{\link{summaryBarPlot}}, \code{\link{backtestBoxPlot}}, 
+#'          \code{\link{backtestChartCumReturn}}, \code{\link{backtestChartStackedBar}}, \code{\link{backtestChartDrawdown}}
+#' 
+#' @examples
+#' \donttest{
+#' library(portfolioBacktest)
+#' data(dataset10)  # load dataset
+#' 
+#' # define your own portfolio function
+#' quintile_portfolio <- function(data, ...) {
+#'   X <- diff(log(data$adjusted))[-1]  
+#'   N <- ncol(X)
+#'   ranking <- sort(colMeans(X), decreasing = TRUE, index.return = TRUE)$ix
+#'   w <- rep(0, N)
+#'   w[ranking[1:round(N/5)]] <- 1/round(N/5)
+#'   return(w)
+#' }
+#' 
+#' # do backtest
+#' bt <- portfolioBacktest(list("Quintile" = quintile_portfolio), dataset10,
+#'                         benchmark = c("1/N", "index"))
+#' 
+#' # now we can chart
+#' backtestChartSharpeRatio(bt)
+#' }
+#' 
+#' @importFrom grDevices topo.colors
+#' @importFrom graphics par
+#' @importFrom PerformanceAnalytics SharpeRatio.annualized chart.RollingPerformance
+#' @importFrom ggplot2 ggplot fortify aes geom_line theme element_blank ggtitle xlab ylab
+#' @importFrom rlang .data
+#' @export
+backtestChartSharpeRatio <- function(bt, portfolios = names(bt), dataset_num = 1, lookback = 100, by = 1, gap = lookback, bars_per_year = 252, type = c("ggplot2", "simple"), ...) {
+  # extract data
+  bt <- bt[portfolios]
+  return <- do.call(cbind, lapply(bt, function(x) x[[dataset_num]]$return))
+  colnames(return) <- names(bt)
+  if (lookback > nrow(return))
+    stop("lookback longer than the time series length!")
+  
+  # plot
+  params <- list(...)
+  switch(match.arg(type),
+         "simple" = {
+           if (is.null(params$col)) params$col <- topo.colors(length(bt))
+           
+           chart.RollingPerformance(return, 
+                                    FUN = function(X) SharpeRatio.annualized(X, scale = bars_per_year, geometric = FALSE), 
+                                    width = lookback, 
+                                    colorset = params$col, lwd = 2, legend.loc = "topleft", 
+                                    main = "Rolling Sharpe Ratio")
+         },
+         "ggplot2" = {
+           # SR_time <- zoo::rollapplyr(return,
+           #                            width = lookback, 
+           #                            FUN = function(X) SharpeRatio.annualized(X, scale = bars_per_year, geometric = FALSE), by.column = TRUE)
+           SR_time <- my_apply_rolling(return, 
+                                       width = lookback, 
+                                       by = by,
+                                       gap = gap,
+                                       FUN = function(X) SharpeRatio.annualized(X, scale = 365*24, geometric = FALSE))
+
+           ggplot(fortify(SR_time, melt = TRUE), aes(x = .data$Index, y = .data$Value, col = .data$Series)) +
+             geom_line() +
+             theme(legend.title = element_blank()) +
+             ggtitle("Rolling Sharpe ratio") + xlab(element_blank()) + ylab("Sharpe ratio")
+         },
+         stop("Unknown plot type."))
+}
+
+my_apply_rolling <- function(R, width = Inf, gap = width, by = 1, FUN = "mean", trim = TRUE, ...) {
+  if (gap == Inf)  # width = Inf is for expanding window
+    gap <- 1
+  res <- c()
+  endings <- seq(from = nrow(R), to = gap, by = -by)
+  endings <- endings[order(endings)]
+  for (ending in endings) {
+    i_start <- max((ending - width + 1), 1)
+    i_end   <- ending
+    res <- rbind(res, 
+                 apply(R[i_start:i_end, ], MARGIN = 2, FUN = FUN, ... = ...))
+  }
+  result_xts <- xts::xts(rbind(NA, res), order.by = zoo::index(R)[c(1, endings)])
+  return(result_xts)
+}
+
+
+
+
+
+
 #' @title Chart of the weight allocation over time for a portfolio over a single backtest
 #' 
 #' @description Create chart of the weight allocation over time for a portfolio over a single 
@@ -445,6 +585,7 @@ backtestChartDrawdown <- function(bt, portfolios = names(bt), dataset_num = 1, t
 #' @param portfolio String with portfolio name to be charted. 
 #'                  Default charts the first portfolio in the backtest.
 #' @param legend Boolean to choose whether legend is plotted or not. Default is \code{legend = FALSE}.
+#' @param num_bars Number of bars shown over time (basically a downsample of the possibly long sequence).
 #' 
 #' @author Daniel P. Palomar and Rui Zhou
 #' 
@@ -484,11 +625,15 @@ backtestChartDrawdown <- function(bt, portfolios = names(bt), dataset_num = 1, t
 #' @importFrom rlang .data
 #' 
 #' @export
-backtestChartStackedBar <- function(bt, portfolio = names(bt[1]), dataset_num = 1, type = c("ggplot2", "simple"), legend = FALSE) {
-  # extract data
-  w <- bt[[portfolio]][[dataset_num]]$w_designed
+backtestChartStackedBar <- function(bt, portfolio = names(bt[1]), dataset_num = 1, num_bars = 100, type = c("ggplot2", "simple"), legend = FALSE) {
   title <- sprintf("Weight allocation over time for %s", portfolio)
-  
+  # extract data and downsample
+  w <- bt[[portfolio]][[dataset_num]]$w_rebalanced
+  w <- w[unique(round(seq(1, nrow(w), length.out = num_bars))), ]
+  w <- w[, colSums(abs(w) > 1e-3) > 0]
+  #w_width <- 1.5 * (as.numeric(last(index(w))) - as.numeric(first(index(w))))/nrow(w)
+  w_width <- max(as.numeric(index(w)[-1] - index(w)[-nrow(w)]))
+
   # plot
   #params <- list(...)
   switch(match.arg(type),
@@ -500,7 +645,7 @@ backtestChartStackedBar <- function(bt, portfolio = names(bt[1]), dataset_num = 
          },
          "ggplot2" = {
            p <- ggplot(fortify(w, melt = TRUE), aes(x = .data$Index, y = .data$Value, fill = .data$Series)) +
-             geom_bar(stat = "identity", color = "black") +
+             geom_bar(stat = "identity", width = w_width) +
              ggtitle(title) + xlab(element_blank()) + ylab("weight")
            if (legend)
              p <- p + labs(fill = "Assets") #theme(legend.title = element_blank())
@@ -510,4 +655,5 @@ backtestChartStackedBar <- function(bt, portfolio = names(bt[1]), dataset_num = 
          },
          stop("Unknown plot type."))
 }
+